+++ ixtb-nextjs/src/lib/endpoints/external/objs/deleteManyObjsEndpoint.ts
@@ -1,7 +1,14 @@
-import { deleteManyObjsSchema, kObjTags } from "fimidx-core/definitions/obj";
+import {
+  deleteManyObjsExternalApiSchema,
+  externalApiQueryToInternalQuery,
+  kObjTags,
+} from "fimidx-core/definitions/obj";
 import { kByTypes } from "fimidx-core/definitions/other";
+import { kFimidxPermissions } from "fimidx-core/definitions/permission";
 import { deleteManyObjs } from "fimidx-core/serverHelpers/index";
+import { checkPermissionProjectThenOrg } from "../../../serverHelpers/permissions";
 import { NextClientTokenAuthenticatedEndpointFn } from "../../types";
+import { sanitizeDeleteManyObjsExternalApiInput } from "../../utils/sanitizeKId0";
 
 export const deleteManyObjsEndpoint: NextClientTokenAuthenticatedEndpointFn<
   void
@@ -11,11 +18,24 @@ export const deleteManyObjsEndpoint: NextClientTokenAuthenticatedEndpointFn<
     session: { clientToken },
   } = params;
 
-  const input = deleteManyObjsSchema.parse(await req.json());
+  const input = deleteManyObjsExternalApiSchema.parse(await req.json());
+  sanitizeDeleteManyObjsExternalApiInput(input);
+
+  await checkPermissionProjectThenOrg({
+    clientToken,
+    projectId: input.projectId,
+    action: kFimidxPermissions.obj.delete,
+  });
+
+  const objQuery = externalApiQueryToInternalQuery(input.query, {
+    projectId: input.projectId,
+    tag: kObjTags.obj,
+  });
+
   await deleteManyObjs({
     deletedBy: clientToken.id,
     deletedByType: kByTypes.clientToken,
-    objQuery: input.query,
+    objQuery,
     tag: kObjTags.obj,
     deleteMany: input.deleteMany ?? false,
   });
