+++ ixtb-nextjs/src/lib/endpoints/internal/orgs/deleteOrgEndpoint.ts
@@ -1,8 +1,10 @@
 import { deleteOrgSchema } from "@/src/definitions/org";
-import { kByTypes } from "fimidx-core/definitions/index";
+import { kByTypes, kFimidxPermissions } from "fimidx-core/definitions/index";
 import { kId0 } from "fimidx-core/definitions/system";
 import { deleteGroups } from "fimidx-core/serverHelpers/index";
+import { requirePermissionForUser } from "../../../serverHelpers/permissions";
 import { NextUserAuthenticatedEndpointFn } from "../../types";
+import { sanitizeDeleteOrgInput } from "../../utils/sanitizeKId0";
 
 export const deleteOrgEndpoint: NextUserAuthenticatedEndpointFn<void> = async (
   params
@@ -16,13 +18,21 @@ export const deleteOrgEndpoint: NextUserAuthenticatedEndpointFn<void> = async (
   const input = deleteOrgSchema.parse({
     id: pathParams.orgId,
   });
+  sanitizeDeleteOrgInput(input);
+
+  await requirePermissionForUser({
+    userId,
+    orgId: input.id,
+    action: kFimidxPermissions.group.delete,
+    target: input.id,
+  });
 
   await deleteGroups({
     query: {
       id: {
         eq: input.id,
       },
-      appId: kId0,
+      projectId: kId0,
     },
     by: userId,
     byType: kByTypes.user,
